T1Ejercicio1

+++ T1E1.cpp
@@ -3,10 +3,10 @@
 using namespace std;
 
 int main(){
-    int num1,num2,num3;
+    float num1,num2,num3;
     cin>>num1>>num2>>num3;
 
-    int promedio=(num1+num2+num3)/3;
+    float promedio=(num1+num2+num3)/3;
 
     cout<< "El resultado del promedio es:" << promedio;
